BrowseDashboards: Split utils
